Move BackendDependencies and IntegrationTestFixture definitions to separate files.
